# Conflicts:
#	lib/mash/presentation/pages/dashboard/parent/parent_dashboard_screen.dart
#	lib/mash/presentation/pages/dashboard/parent/widget/parent_dashboard_student_detail_widget.dart
#	lib/mash/presentation/pages/dashboard/parent/widget/parent_dashboard_top_section.dart
#	pubspec.yaml

+++ lib/mash/presentation/pages/dashboard/parent/widget/dashboard_common_last_section.dart
@@ -141,7 +141,7 @@ class DashboardCommonLastWidget extends StatelessWidget {
                   final isWord = data!.resTable1!.length > index;
                   int tempIndex = 0;
                   if (!isWord) {
-                    tempIndex = index - data.resTable1!.length;
+                    tempIndex = (index - data.resTable1!.length).toInt();
                   }
                   return !isWord
                       ? wordImageWidget(
